requirements.txt
Flask
opencv-python
pytesseract
pillow
